+++ Client/.eslintignore
@@ -1,0 +1,4 @@
+docs/
+public/
+dist/
+node_modules/
