+++ Client/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "ACSCallingSample",
+  "version": "0.3.0",
+  "private": true,
+  "dependencies": {
+    "@azure/communication-administration": "1.0.0-beta.4",
+    "@azure/communication-calling": "1.6.2-beta.2",
+    "@fluentui/react": "^7.158.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0"
+  },
+  "scripts": {
+    "build": "npm run clean && concurrently --kill-others-on-fail \"tsc\" \"webpack --mode=production --env production\"",
+    "clean": "rimraf dist",
+    "package": "copyfiles -E -u 3 \"../Server/dist/**/*\" dist && rm -r dist/dist-esm",
+    "start": "concurrently \"cd ../Server && npm run start:dev\" \"npm run start:app\"",
+    "start:app": "webpack-dev-server --port 5000 --mode development",
+    "test": "",
+    "prettier": "prettier --no-error-on-unmatched-pattern --write --config ./.prettierrc --ignore-path=../.prettierignore \"**/*.js\" \"**/*.jsx\"",
+    "prettier:check": "prettier --no-error-on-unmatched-pattern --check --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\"",
+    "lint": "eslint */**/*.{js,jsx}",
+    "lint:fix": "eslint */**/*.{js,jsx} --fix",
+    "lint:quiet": "eslint */**/*.{js,jsx} --quiet"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.8.7",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.7",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/runtime": "^7.8.7",
+    "babel-loader": "^8.0.6",
+    "concurrently": "^5.3.0",
+    "copyfiles": "^2.4.1",
+    "cross-env": "^7.0.3",
+    "css-loader": "^5.2.6",
+    "eslint": "^8.20.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-header": "^3.1.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "style-loader": "^1.1.3",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.11.2"
+  },
+  "peerDependencies": {
+    "@azure/logger": "^1.0.2"
+  }
+}
