--- Project/config.json
@@ -1,5 +1,0 @@
-{
-    "connectionString": "REPLACE_WITH_CONNECTION_STRING"
-}
-
-
